Move ipc-Handlers from main to specific file
Just for simplicity in main.js
